examples/simple/python_client/requirements.txt
# Official A2A SDK dependencies  
a2a-sdk>=0.1.0
httpx>=0.24.0 